requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
line-bot-sdk==3.12.0
openai>=1.58.0
google-auth==2.29.0
python-dotenv==1.0.1
langfuse>=2.57.0
